+++ R/plot_daily_stats.R
@@ -15,19 +15,21 @@
 #' @description Plots means, medians, maximums, minimums, and percentiles for each day of the year of flow values 
 #'    from a daily streamflow data set. Can determine statistics of rolling mean days (e.g. 7-day flows) using the \code{roll_days} 
 #'    argument. Calculates statistics from all values, unless specified. The Maximum-Minimum band can be removed using the 
-#'    \code{include_extremes} argument and the percentile bands can be customized using the \code{inner_percentiles} and 
+#'    \code{plot_extremes} argument and the percentile bands can be customized using the \code{inner_percentiles} and 
 #'    \code{outer_percentiles} arguments. Data calculated using \code{calc_daily_stats()} function. Returns a list of plots.
 #'
 #' @inheritParams calc_daily_stats
 #' @inheritParams plot_annual_stats
 #' @param add_year Numeric value indicating a year of daily flows to add to the daily statistics plot.  Leave blank
 #'    or set to \code{NULL} for no years.
-#' @param include_extremes Logical value to indicate plotting a ribbon with the range of daily minimum and maximum flows. 
+#' @param plot_extremes Logical value to indicate plotting a ribbon with the range of daily minimum and maximum flows. 
 #'    Default \code{TRUE}.
 #' @param inner_percentiles Numeric vector of two percentile values indicating the lower and upper limits of the 
 #'    inner percentiles ribbon for plotting. Default \code{c(25,75)}, set to \code{NULL} for no inner ribbon.
 #' @param outer_percentiles Numeric vector of two percentile values indicating the lower and upper limits of the 
 #'    outer percentiles ribbon for plotting. Default \code{c(5,95)}, set to \code{NULL} for no outer ribbon.
+#' @param plot_inner_percentiles Logical value indicating whether to plot the inner percentiles ribbon. Default \code{TRUE}.
+#' @param plot_outer_percentiles Logical value indicating whether to plot the outer percentiles ribbon. Default \code{TRUE}.
 #'
 #' @return A list of ggplot2 objects with the following for each station provided:
 #'   \item{Daily_Stats}{a plot that contains daily flow statistics}
@@ -83,7 +85,9 @@ plot_daily_stats <- function(data,
                              complete_years = FALSE,
                              months = 1:12,
                              ignore_missing = FALSE,
-                             include_extremes = TRUE,
+                             plot_extremes = TRUE,
+                             plot_inner_percentiles = TRUE,
+                             plot_outer_percentiles = TRUE,
                              inner_percentiles = c(25,75),
                              outer_percentiles = c(5,95),
                              add_year,
@@ -120,7 +124,9 @@ plot_daily_stats <- function(data,
   add_year_checks(add_year)
   logical_arg_check(include_title)
   ptile_ribbons_checks(inner_percentiles, outer_percentiles)
-  
+  logical_arg_check(plot_extremes)
+  logical_arg_check(plot_inner_percentiles)
+  logical_arg_check(plot_outer_percentiles)
   
   ## FLOW DATA CHECKS AND FORMATTING
   ## -------------------------------
@@ -140,7 +146,7 @@ plot_daily_stats <- function(data,
   origin_date <- get_origin_date(water_year_start)
   
   
-
+  
   ## CALC STATS
   ## ----------
   
@@ -203,7 +209,7 @@ plot_daily_stats <- function(data,
   # Create manual colour and fill options
   
   fill_manual_list <- c()
-  if (include_extremes) {
+  if (plot_extremes) {
     fill_manual_list <- c(fill_manual_list, "lightblue2")
     names(fill_manual_list) <- c(names(fill_manual_list), "Minimum-Maximum")
   }
@@ -243,13 +249,15 @@ plot_daily_stats <- function(data,
     plot = purrr::map2(
       data, STATION_NUMBER,
       ~ggplot2::ggplot(data = ., ggplot2::aes(x = Date)) +
-        {if(include_extremes) ggplot2::geom_ribbon(ggplot2::aes(ymin = Minimum, ymax = Maximum, fill = "Minimum-Maximum"), na.rm = FALSE)} +
-        {if(is.numeric(outer_percentiles)) ggplot2::geom_ribbon(ggplot2::aes_string(ymin = paste0("P",min(outer_percentiles)),
-                                                                                    ymax = paste0("P",max(outer_percentiles)),
-                                                                                    fill = paste0("'",outer_name,"'")), na.rm = FALSE)} +
-        {if(is.numeric(inner_percentiles)) ggplot2::geom_ribbon(ggplot2::aes_string(ymin = paste0("P",min(inner_percentiles)),
-                                                                                    ymax = paste0("P",max(inner_percentiles)),
-                                                                                    fill = paste0("'",inner_name,"'")), na.rm = FALSE)} +
+        {if(plot_extremes) ggplot2::geom_ribbon(ggplot2::aes(ymin = Minimum, ymax = Maximum, fill = "Minimum-Maximum"), na.rm = FALSE)} +
+        {if(is.numeric(outer_percentiles) & plot_outer_percentiles) 
+          ggplot2::geom_ribbon(ggplot2::aes_string(ymin = paste0("P",min(outer_percentiles)),
+                                                   ymax = paste0("P",max(outer_percentiles)),
+                                                   fill = paste0("'",outer_name,"'")), na.rm = FALSE)} +
+        {if(is.numeric(inner_percentiles) & plot_inner_percentiles)
+          ggplot2::geom_ribbon(ggplot2::aes_string(ymin = paste0("P",min(inner_percentiles)),
+                                                   ymax = paste0("P",max(inner_percentiles)),
+                                                   fill = paste0("'",inner_name,"'")), na.rm = FALSE)} +
         ggplot2::geom_line(ggplot2::aes(y = Median, colour = "Median"), size = .5, na.rm = TRUE) +
         ggplot2::geom_line(ggplot2::aes(y = Mean, colour = "Mean"), size = .5, na.rm = TRUE) +
         {if(!log_discharge) ggplot2::scale_y_continuous(expand = c(0, 0), breaks = scales::pretty_breaks(n = 8),